chore: apply changes requested in structure
