+++ src/dfe/nfe/webServices/requests/common.ts
@@ -1,0 +1,10 @@
+export interface NfeRequestOptions<Body> {
+  body: Body;
+  timeout: number;
+}
+
+export interface NfeWebServiceResponse<Status, Raw> {
+  status: Status | "outro";
+  description: string;
+  raw: Raw;
+}
