+++ src/dfe/nfe/webServices/requests/index.ts
@@ -1,0 +1,15 @@
+export type { NfeRequestOptions, NfeWebServiceResponse } from "./common";
+
+export type {
+  NfeConsultaCadastroOptions,
+  NfeConsultaCadastroStatus,
+  NfeConsultaCadastroResponseRaw,
+  NfeConsultaCadastroRequest,
+  NfeConsultaCadastroResponse,
+} from "./consultaCadastro";
+export type {
+  NfeStatusServicoStatus,
+  NfeStatusServicoResponseRaw,
+  NfeStatusServicoRequest,
+  NfeStatusServicoResponse,
+} from "./statusServico";
