+++ src/ufCode/types.ts
@@ -1,0 +1,64 @@
+import type { LiteralStringUnion } from "@/utils";
+
+export type UF = LiteralStringUnion<
+  | "AC"
+  | "AL"
+  | "AP"
+  | "AM"
+  | "BA"
+  | "CE"
+  | "DF"
+  | "ES"
+  | "GO"
+  | "MA"
+  | "MT"
+  | "MS"
+  | "MG"
+  | "PA"
+  | "PB"
+  | "PR"
+  | "PE"
+  | "PI"
+  | "RJ"
+  | "RN"
+  | "RS"
+  | "RO"
+  | "RR"
+  | "SC"
+  | "SP"
+  | "SE"
+  | "TO"
+>;
+
+/**
+ * @description Código numérico de uma UF
+ */
+export type UFCode = LiteralStringUnion<
+  | "12"
+  | "27"
+  | "16"
+  | "13"
+  | "29"
+  | "23"
+  | "53"
+  | "32"
+  | "52"
+  | "21"
+  | "51"
+  | "50"
+  | "31"
+  | "15"
+  | "25"
+  | "41"
+  | "26"
+  | "22"
+  | "33"
+  | "24"
+  | "43"
+  | "11"
+  | "14"
+  | "42"
+  | "35"
+  | "28"
+  | "17"
+>;
